Add ability for IRC bots to echo each other's output.
Draft of inter-server communication using IRC bots.

Two IRC bots in the same IRC channel (and on different servers) will be
able to output the broadcasts and deaths from one server to another.
E.g. if Server103Chat echoes a broadcast from server 103, Server105Chat
can convert it back to a m59-compatible string and broadcast it on 105,
using the dm gqemote command to avoid feedback loops.

For a bot to echo other bots' text, it will need to list that bot's name
and a banner (for when the text hits the bots own server) in
configuration.xml.